+++ buildSrc/src/main/kotlin/BuildConfig.kt
@@ -3,8 +3,8 @@ package com.unciv.build
 object BuildConfig {
     const val kotlinVersion = "1.3.71"
     const val appName = "Unciv"
-    const val appCodeNumber = 431
-    const val appVersion = "3.8.12"
+    const val appCodeNumber = 437
+    const val appVersion = "3.9.2"
 
     const val gdxVersion = "1.9.10"
     const val roboVMVersion = "2.3.1"