+++ core/src/com/unciv/logic/battle/BattleDamage.kt
@@ -110,23 +110,23 @@ object BattleDamage {
         return modifiers
     }
 
-    fun getAttackModifiers(attacker: ICombatant, defender: ICombatant): HashMap<String, Float> {
+    fun getAttackModifiers(attacker: ICombatant, tileToAttackFrom:TileInfo?, defender: ICombatant): HashMap<String, Float> {
         val modifiers = getGeneralModifiers(attacker, defender)
         val policies = attacker.getCivInfo().policies
 
-        if(attacker is MapUnitCombatant) {
-            modifiers.putAll(getTileSpecificModifiers(attacker,defender.getTile()))
+        if (attacker is MapUnitCombatant) {
+            modifiers.putAll(getTileSpecificModifiers(attacker, defender.getTile()))
 
             for (ability in attacker.unit.getUniques()) {
                 val regexResult = Regex(BONUS_AS_ATTACKER).matchEntire(ability) //to do: extend to defender, and penalyy
                 if (regexResult == null) continue
                 val bonus = regexResult.groups[1]!!.value.toFloat() / 100
                 if (modifiers.containsKey("Attacker Bonus"))
-                    modifiers["Attacker Bonus"] =modifiers["Attacker Bonus"]!! + bonus
+                    modifiers["Attacker Bonus"] = modifiers["Attacker Bonus"]!! + bonus
                 else modifiers["Attacker Bonus"] = bonus
             }
 
-            if(attacker.unit.isEmbarked() && !attacker.unit.hasUnique("Amphibious"))
+            if (attacker.unit.isEmbarked() && !attacker.unit.hasUnique("Amphibious"))
                 modifiers["Landing"] = -0.5f
 
             if (attacker.isMelee()) {
@@ -136,7 +136,15 @@ object BattleDamage {
                             && MapUnitCombatant(it.militaryUnit!!).isMelee()
                 }
                 if (numberOfAttackersSurroundingDefender > 1)
-                    modifiers["Flanking"] = 0.1f * (numberOfAttackersSurroundingDefender-1) //https://www.carlsguides.com/strategy/civilization5/war/combatbonuses.php
+                    modifiers["Flanking"] = 0.1f * (numberOfAttackersSurroundingDefender - 1) //https://www.carlsguides.com/strategy/civilization5/war/combatbonuses.php
+
+                if (tileToAttackFrom != null && tileToAttackFrom.isConnectedByRiver(defender.getTile())) {
+                    if (!tileToAttackFrom.hasConnection(attacker.getCivInfo()) // meaning, the tiles are not road-connected for this civ
+                            || !defender.getTile().hasConnection(attacker.getCivInfo())
+                            || !attacker.getCivInfo().tech.roadsConnectAcrossRivers) {
+                        modifiers["Across river"] = -0.2f
+                    }
+                }
             }
 
             if (policies.autocracyCompletedTurns > 0 && policies.hasEffect("+20% attack bonus to all Military Units for 30 turns"))
@@ -145,15 +153,11 @@ object BattleDamage {
             if (defender is CityCombatant &&
                     attacker.getCivInfo().containsBuildingUnique("+15% Combat Strength for all units when attacking Cities"))
                 modifiers["Statue of Zeus"] = 0.15f
-        }
-
-        else if (attacker is CityCombatant) {
+        } else if (attacker is CityCombatant) {
             if (policies.hasEffect("Units in cities cost no Maintenance, garrisoned city +50% attacking strength")
                     && attacker.city.getCenterTile().militaryUnit != null)
                 modifiers["Oligarchy"] = 0.5f
         }
-
-
 
         return modifiers
     }
@@ -217,7 +221,8 @@ object BattleDamage {
                         || tile.terrainFeature != Constants.jungle))
             modifiers[tile.baseTerrain] = 0.25f
 
-        if(unit.getCivInfo().nation.unique == UniqueAbility.WAYFINDING && tile.getTilesInDistance(2).any { it.improvement=="Moai" })
+        if(unit.getCivInfo().nation.unique == UniqueAbility.WAYFINDING
+                && tile.getTilesInDistance(2).any { it.improvement=="Moai" })
             modifiers["Moai"] = 0.1f
 
         if(tile.neighbors.flatMap { it.getUnits() }
@@ -265,8 +270,8 @@ object BattleDamage {
     /**
      * Includes attack modifiers
      */
-    private fun getAttackingStrength(attacker: ICombatant, defender: ICombatant): Float {
-        val attackModifier = modifiersToMultiplicationBonus(getAttackModifiers(attacker,defender))
+    private fun getAttackingStrength(attacker: ICombatant, tileToAttackFrom: TileInfo?, defender: ICombatant): Float {
+        val attackModifier = modifiersToMultiplicationBonus(getAttackModifiers(attacker,tileToAttackFrom, defender))
         return attacker.getAttackingStrength() * attackModifier
     }
 
@@ -280,15 +285,15 @@ object BattleDamage {
         return defender.getDefendingStrength() * defenceModifier
     }
 
-    fun calculateDamageToAttacker(attacker: ICombatant, defender: ICombatant): Int {
+    fun calculateDamageToAttacker(attacker: ICombatant, tileToAttackFrom: TileInfo?, defender: ICombatant): Int {
         if(attacker.isRanged()) return 0
         if(defender.getUnitType().isCivilian()) return 0
-        val ratio = getAttackingStrength(attacker,defender) / getDefendingStrength(attacker,defender)
+        val ratio = getAttackingStrength(attacker, tileToAttackFrom, defender) / getDefendingStrength(attacker,defender)
         return (damageModifier(ratio, true) * getHealthDependantDamageRatio(defender)).roundToInt()
     }
 
-    fun calculateDamageToDefender(attacker: ICombatant, defender: ICombatant): Int {
-        val ratio = getAttackingStrength(attacker,defender) / getDefendingStrength(attacker,defender)
+    fun calculateDamageToDefender(attacker: ICombatant, tileToAttackFrom: TileInfo?, defender: ICombatant): Int {
+        val ratio = getAttackingStrength(attacker,tileToAttackFrom, defender) / getDefendingStrength(attacker,defender)
         return (damageModifier(ratio,false) * getHealthDependantDamageRatio(attacker)).roundToInt()
     }