+++ core/src/com/unciv/logic/map/TileInfo.kt
@@ -341,12 +341,15 @@ open class TileInfo {
         return toString(null)
     }
 
+    /** The two tiles have a river between them */
     fun isConnectedByRiver(otherTile:TileInfo): Boolean {
-        if(otherTile !in neighbors) throw Exception("Should never call this function on a non-neighbor!")
+        if(otherTile==this) throw Exception("Should not be called to compare to self!")
         val xDifference = this.position.x - otherTile.position.x
         val yDifference = this.position.y - otherTile.position.y
 
         return when {
+            yDifference < -1f || xDifference < -1f || yDifference > 1f || xDifference > 1f ->
+                throw Exception("Should never call this function on a non-neighbor!")
             xDifference == 1f && yDifference == 1f -> hasBottomRiver // we're directly above it
             xDifference == 1f -> hasBottomRightRiver // we're to the top-left of it
             yDifference == 1f -> hasBottomLeftRiver // we're to the top-right of it
@@ -450,9 +453,6 @@ open class TileInfo {
 
     fun hasConnection(civInfo: CivilizationInfo) =
             roadStatus != RoadStatus.None || forestOrJungleAreRoads(civInfo)
-
-    fun hasRoad(civInfo: CivilizationInfo) =
-            roadStatus == RoadStatus.Road || forestOrJungleAreRoads(civInfo)
 
     fun hasRailroad() =
             roadStatus == RoadStatus.Railroad