+++ core/src/com/unciv/ui/worldscreen/WorldScreenTopBar.kt
@@ -120,8 +120,10 @@ class WorldScreenTopBar(val worldScreen: WorldScreen) : Table() {
                 .apply { setSize(50f, 50f) }
         menuButton.color = Color.WHITE
         menuButton.onClick {
-            if (worldScreen.popups.none { it is WorldScreenMenuPopup })
-                WorldScreenMenuPopup(worldScreen).open(force = true)
+            val worldScreenMenuPopup = worldScreen.popups.firstOrNull { it is WorldScreenMenuPopup }
+            if(worldScreenMenuPopup!=null)
+                worldScreenMenuPopup.close()
+            else WorldScreenMenuPopup(worldScreen).open(force = true)
         }
         menuButton.centerY(this)
         menuButton.x = menuButton.y