+++ changelog.md
@@ -1,3 +1,43 @@
+## 3.9.2
+
+Move units out of cities when liberating
+
+Thread crash fixes - By vainiovano
+
+New translation language - Lithuanian
+
+Clicking the menu button when it is open closes the menu
+
+Translation updates
+
+## 3.9.1
+
+Buildings requiring a nearby resource can be constructed even when the tile belongs to another city
+
+Natural wonders spawned before rivers, so we don't retroactively get rivers on coast tiles
+
+Added civilopedia info for great people and great improvements, removing terrain features, and strategic resource provision - see #1492
+
+Resolved #2613 - added a close button to the Civ-picking popup in the New Game screen
+
+Battle calculation takes into account the tile that the unit will attack from
+
+Translation updates
+
+## 3.9.0
+
+Added rivers, and river generation!
+
+Game now saves save files in external storage on Android when possible.
+
+Resolved #2672 - Difficulties are sorted by ascending difficulty in Civilopedia
+
+Resolved #2660 - Remove Fallout now enabled by Atomic Theory, not Agriculture
+
+Great person uniques can be added to any modded unit
+
+Translation updates
+
 ## 3.8.12
 
 By dbaelz: